Add DeleteRoom Modal

+++ src/MrHotel/WebApp/src/app/pages/reservations/reservations.page.ts
@@ -9,10 +9,10 @@ import { Observable } from "rxjs";
 import { RoomStatusComponent } from "@components/room-status/room-status.component";
 import { ModalController } from "@ionic/angular";
 import { AddRoomModalFormComponent } from "@components/modals/add-room-modal-form";
+import { DeleteRoomModalFormComponent } from "@components/modals/delete-room-modal-form";
 import { ReservationManagerService } from "@services/reservation-manager";
 import { RoomStatus, Room } from "@services/reservation-manager/data";
 import { IonicModule } from "@ionic/angular";
-import * as Random from "@utilities/random";
 import { BaseModalFormComponent } from "@components/modals/modal-base-form-component";
 
 @Component({
@@ -25,6 +25,7 @@ import { BaseModalFormComponent } from "@components/modals/modal-base-form-compo
     RoomStatusComponent,
     AddRoomModalFormComponent,
     IonicModule,
+    DeleteRoomModalFormComponent
   ],
 })
 export class ReservationsPage {
@@ -40,23 +41,28 @@ export class ReservationsPage {
   }
 
   public async addRoom(): Promise<void> {
-    const roomData = await this.openModal(AddRoomModalFormComponent);
+    const roomData = await this.openModal(AddRoomModalFormComponent, "add-room-modal");
 
-    if(roomData) {
+    if (roomData.name) {
       this.reservationManager.addRoom(roomData.name).subscribe();
     }
   }
 
-  public deleteRoom(room: Room): void {
-    this.reservationManager.deleteRoom(room).subscribe();
+  public async deleteRoom(room: Room): Promise<void> {
+    const isDeleteConfirmed = await this.openModal(DeleteRoomModalFormComponent, "delete-room-modal");
+
+    if (isDeleteConfirmed.state){
+      this.reservationManager.deleteRoom(room).subscribe();
+    }
   }
 
   private async openModal(
     component: Type<BaseModalFormComponent>,
+    cssClass: string,
   ): Promise<any> {
     const modal = await this.modalController.create({
       component: component,
-      cssClass: 'add-room-modal'
+      cssClass: cssClass,
     });
 
     await modal.present();
